Add logarithmic scaling Add windowing function (Hamming) Add message support (top right corner) Change keymap

+++ apps/plugins/fft/math.h
@@ -7,7 +7,7 @@
 #define Q_MUL(a, b, bits) (( (int64_t) (a) * (int64_t) (b) ) >> (bits))
 #define Q15_MUL(a, b) Q_MUL(a,b,15)
 
-#define Q_DIV(a, b, bits) (int32_t)( (((int64_t) (a)) << (bits)) / (b) )
+#define Q_DIV(a, b, bits) ( (((int64_t) (a)) << (bits)) / (b) )
 #define Q15_DIV(a, b) Q_DIV(a,b,15)
 
 #define float_q(a, bits) (int32_t)( ((float)(a)) *(1<<(bits)))
